Everything looks good on Linux now.
Also, the weird projection bug IS related to the temp objects or files not being cleared properly!!!!!! Now we jsut need to track it down!!!

AuthorForm => Add Color, Fix Top Label Size

Info => XML converter does not set a default noDataValue

ResampleForm => Add Color, Make Default Action (enter pressed) the Ok button

I'll be looking at all these when I get back to Windows

+++ mapimg/mapimg.cpp
@@ -1,4 +1,4 @@
-// $Id: mapimg.cpp,v 1.10 2005/02/10 15:48:19 jtrent Exp $
+// $Id: mapimg.cpp,v 1.11 2005/02/10 16:59:12 jtrent Exp $
 
 
 #include "mapimg.h"
@@ -17,10 +17,10 @@ extern "C"
 }
 
 #include <math.h>
-double mapimg::round(double value, unsigned int decimals)
+int mapimg::round(double value, unsigned int decimals)
 {
   double factor = pow(10,decimals);
-  return floor((value * factor) + 0.5) / factor;
+  return (int)(floor((value * factor) + 0.5) / factor);
 }
 
 bool mapimg::readytoFrameIt( RasterInfo &input, QWidget * parent )
